Restart vncserver if gnome-session fails

+++ vncserver/start-template-v3.sh
@@ -16,16 +16,22 @@ echo "scancel ${SLURM_JOB_ID}"  >> ${resource_jobdir}/cancel.sh
 start_gnome_session_with_retries() {
     k=1
     while true; do
-        if [ $k -gt 1 ]; then
-            echo "$(date) Restarting vncserver"
-            ${service_vnc_exec} -kill ${DISPLAY}
-            sleep 3
-            ${service_vnc_exec} ${DISPLAY} -SecurityTypes VncAuth -PasswordFile ${resource_jobdir}/.vncpasswd
+        if xset q >/dev/null 2>&1; then
+            echo "(date) X server on $DISPLAY is alive."
+            sleep $((k*10))
+        else
+            echo "(date) X server on $DISPLAY is unresponsive."
+            if [ $k -gt 1 ]; then
+                echo "$(date) Restarting vncserver"
+                ${service_vnc_exec} -kill ${DISPLAY}
+                sleep 3
+                ${service_vnc_exec} ${DISPLAY} -SecurityTypes VncAuth -PasswordFile ${resource_jobdir}/.vncpasswd
+            fi
+            sleep 2
+            echo "$(date) Starting gnome-session"
+            gnome-session --debug
+            sleep $((k*10))
         fi
-        sleep 2
-        echo "$(date) Starting gnome-session"
-        gnome-session
-        sleep $((k*10))
         k=$((k+1))
     done
 }
